added comments to rng_wrapper

+++ src/multi_physics/QED/src/utils/rng_wrapper.hpp
@@ -1,29 +1,46 @@
 #ifndef PICSAR_MULTIPHYSICS_RNG_WRAPPER
 #define PICSAR_MULTIPHYSICS_RNG_WRAPPER
 
-//This .hpp file provides a wrapper for the RNG.
-//It builds a wrapper around the 64 bit Mersenne Twister availabe in the STL.
+//Should be included by all the src files of the library
+#include "../qed_commons.h"
 
 #include <random>
 #include <cmath>
 #include <cstdint>
 
-//Should be included by all the src files of the library
-#include "../qed_commons.h"
-
 namespace picsar{
 namespace multi_physics{
 namespace utils{
 
-    //*** STL RNG Wrapper ***
+    /**
+    * This class is a wrapper around the 64 bit Mersenne Twister generator
+    * and the distribution functions provided by the standard template library.
+    * It can be used to instantiate the engine objects of the high order interface
+    * of the library. However, any other used-defined class implementing the
+    * same interface can be used as well.
+    */
     class stl_rng_wrapper
     {
     public:
-        //The seed must be a 64 bit integer
+
+        /**
+        * The constructor initializes the random number generator
+        *
+        * @param[in] seed a 64bit integer
+        */
         stl_rng_wrapper(uint64_t seed):
             m_rng{seed}{}
 
-        //Get rnd number uniformly distributed in [a,b)
+
+        /**
+        * This function returns floating point numbers uniformly
+        * distributed in [a,b).
+        *
+        * @tparam RealType the floating point number type
+        * @param[in] a the lower limit of the interval
+        * @param[in] b the upper limit of the interval
+        * @return a random number uniformly distributed in [a,b)
+        */
         template<typename RealType>
         PXRMP_INTERNAL_FORCE_INLINE_DECORATOR
         RealType unf(RealType a, RealType b)
@@ -33,7 +50,14 @@ namespace utils{
             return unf_dist_a_b(m_rng);
         }
 
-        //Get rnd number with exponential distribution
+        /**
+        * This function returns positive floating point numbers with
+        * an exponential probability distribution f(x) = l * exp(-l * x)
+        *
+        * @tparam RealType the floating point number type
+        * @param[in] l the parameter of the exponential distribution
+        * @return a random number with an exponential distribution
+        */
         template<typename RealType>
         PXRMP_INTERNAL_FORCE_INLINE_DECORATOR
         RealType exp(RealType l)
@@ -43,17 +67,32 @@ namespace utils{
             return exp_dist_l(m_rng);
         }
 
-        //Get rnd number with poisson distribution
+        /**
+        * This function returns positive 64bit integers with
+        * a Poisson's distribution (with parameter l)
+        *
+        * @tparam RealType the floating point number type of the parameter l
+        * @param[in] l the parameter of the Poisson's distribution
+        * @return a random integer with a Poisson's distribution
+        */
         template<typename RealType>
         PXRMP_INTERNAL_FORCE_INLINE_DECORATOR
-        size_t poisson(RealType l)
+        uint64_t poisson(RealType l)
         {
             auto poisson_dist_l =
-                std::poisson_distribution<size_t>(l);
+                std::poisson_distribution<uint64_t>(l);
             return poisson_dist_l(m_rng);
         }
 
-        //Get rnd number with gaussian distribution
+        /**
+        * This function returns floating point numbers with
+        * a gaussian distribution (with given mean and deviation)
+        *
+        * @tparam RealType the floating point number type
+        * @param[in] mean the mean of the gaussian distribution
+        * @param[in] deviation the standard deviation of the gaussian distribution
+        * @return a random number with a gaussian distribution
+        */
         template<typename RealType>
         PXRMP_INTERNAL_FORCE_INLINE_DECORATOR
         RealType gaussian(RealType mean, RealType deviation)
@@ -64,7 +103,6 @@ namespace utils{
         }
 
     private:
-        //Internally the 64 bit Mersenne Twister generator is used
         std::mt19937_64 m_rng;
 
     };
